Add eth-infinitism/account-abstraction submodule
Add eth-infinitism's repo with the reference contracts as a submodule.
We'll be using these contracts for a few things:

* Generating the ABIs for `EntryPoint` and (for testing) `IAccount`,
  etc.
* Deploying the `EntryPoint` to a local Geth node during testing.
* As imports for our own contracts that we use in tests (like custom
  wallets, etc.).

Also adds a script to generate the `EntryPoint` ABI programmatically
rather than having it static in our project. Call this script in
`build.rs` (but only if the ABI doesn't exist yet. Otherwise it adds too
much time to startup).