Rust side of simulation logic
An untested, somewhat messy, yet possibly complete implementation of
simulation logic on the Rust side. At the moment, the tracer hasn't been
written so we can't try this yet, but the Rust logic for checking the
returned value is ready.

Notably, we return a list of violations on failure, which differs from
the reference implementation which only returns the first failure. We
may want to keep an eye on this to avoid DoS's (we can always
short-circuit after a certain number of failures).