Represent chain id as a u64
Ethers uses `u64` to represent chain ids, so there's no point having the
rest of our code represent it with `U256`. It also simplifies slightly,
because we can use a `uint64` in the protobuf definition instead of
needing to convert to/from `bytes`, and reduces the number of `.into()`s
needed.