Add ValidTimeRange type
Create a new type, `ValidTimeRange`, which represents the `(valid_after,
valid_before)` pairs returned by wallets and paymasters. This helps
ensure we use the correct range-checking logic each time we use this,
including the easy-to-miss rule that a `valid_before` of zero means that
the end time is not bounded.