Allow EntryPoint to make arbitrary accesses
Accesses made by the entry point (at the top-level) should not count
against the operation when calculating violations.

To implement this, refactor the tracer return format a little bit, since
we still need to keep track of entry point accesses to compute expected
values for conditional transactions, but we no longer want them to count
for violations.