Fix build.rs always rerunning
A typo in `build.rs` was causing it to re-run on every build instead of
only when the files it depends on change.

My guess for why this happens: Tonic's build script emits a
"rerun-if-changed" for its "includes" directory. Here the "includes"
directory was set to "builder" which doesn't exist, and emitting
"rerun-if-changed" for a nonexistent file causes it to rerun every time.