feat(builder): Add `BundleProposer`
Adds a new trait `BundleProposer` and its implementation for assembling
a bundle on demand. The proposer retrieves ops from the mempool, runs
simulation on them, and then marks them for inclusion or rejection
appropriately. It also groups ops by their aggregators and manages the
aggregator signatures appropriately.

This does not yet actually send the ops in a transaction. That can be
the responsibility of a separate struct.

A significant amount of the work here went into setting up the unit
tests, but the testing is worth it because of how complicated some of
the logic can get. One particular test improvement is that the gRPC
mocks will now search for an open port instead of just trying a single
one- the old behavior ran into port conflicts when running tests in
parallel.