feat(builder): Run `eth_estimateGas` when proposing a bundle
Updates the bundle proposer logic so that after putting together a
bundle, it runs `eth_estimateGas` calling `handleOps` (or
`handleAggregatedOps`). If successful, it inclues the gas estimate in
the returned bundle. If one of the ops or aggregators fails, then reject
that op and try again.

This turns out to be incredibly painful to write. To do so, refactor the
bundle proposer logic so that it maintains a `ProposerContext` struct
that represents the current state what is needed to assemble a bundle.
This context has various helper methods such as for rejecting the op
with a specified index, and is mutated as the proposer iterates on the
bundle until it finds one that succeeds.

Also introduce a new mockable trait, `EntryPointLike`, which exposes
convenience methods for what we'll need from an `IEntryPoint`, then
update the tests to use it.