feat(builder): Implement builder
Implements the builder, which looks to the op pool on every new block to
try to form a bundle. If it does, it sends the bundle in a transaction,
then waits for that transaction to become mined or dropped before
attempting any additional bundles.

Various refactorings were needed to support this. Most notably,
introduce a new `BundlerSigner` enum which allows us to produce a
`Signer` backed by either a local signer or a KMS signer, selected at
runtime. This allows us to use an `IEntryPoint` backed by this signer to
send our transactions, and we do so behind the `EntryPointLike` trait.