fix(rpc): Account for updated op when estimating pre-verification gas
While the current pre-verification gas estimate is technically correct
for an op, it becomes incorrect if the user takes this estimate and
plugs it into the op because the estimate depends on the number of
nonzero bytes in the packed op.

Instead, pessimistically assume that all gas estimates will be entirely
nonzero bytes, to be sure that the pre-verification gas will be enough
after updating the op.