feat(rpc): Use binary search for gas estimation
Replaces the current gas estimation, which uses a tracer and returns the
difference between starting gas and ending gas, with a new version that
uses binary search instead. This is more accurate (and in fact can be
exact) and avoids all kinds of tracer woes on different networks.

We work around the problem of the entry point's `simulateHandleOp`
method not indicating whether the call portion succeded by creating a
proxy contract that will perform the call gas estimation's binary search
and will be called via `eth_call` overrides. For a more detailed
explanation, see the comment at the top of `CallGasEstimationProxy.sol`.