feat(builder): Choose priority fees to not lose money
If left to its own devices, ethers-rs always sets
`max_priority_fee_per_gas` to at least 3 gwei when sending transactions,
which is rather high (transactions on eth-mainnet often set this to 0.1
gwei or even 0). This was losing us money, as we were accepting
user operations as long as their `max_priority_fee_per_gas` was at least
0.1 gwei.

Instead, call `eth_maxPriorityFeePerGas` to get a recommended value for
this field, and one which users sending ops can also compute simply. We
then only include ops in bundles if their priority fee is at least 10%
higher than this field (configurable). Note that we still also have the
old requirement that ops must have a priority fee of at least 0.1 gwei
before they are even added to the op pool.

Be careful around Optimism mainnet, which doesn't support this method at
the moment. Don't call it if the chain id is Opt mainnet or any Arbitrum
chain (which doesn't use this field), and also add a new setting where
users can override this behavior (for example, if they are running a
local Optimism node with a custom chain id and don't want this method
being called).