fix(estimation): enforce MIN_CALL_GAS_LIMIT in estimation
When returning estimates for `min_call_gas`, ensure that the returned
value is at least `MIN_CALL_GAS_LIMIT`, a value mandated by the spec
below which the precheck should reject the operation.

While we're at it, update `MIN_CALL_GAS_LIMIT` from 9000 to 9100. The
spec says it should be the minimum cost of a call with nonzero value,
which is 9100.