fix(estimation): use positive gas fee in validation estimation
When `maxFeePerGas` is zero, the account's required prepayment is zero
as well, which means the account skips an eth transfer to the entry
point. This missing transfer means that gas estimates are too low if the
user sets their gas fees after performing gas estimation. To avoid this,
always do verification gas estimation with a positive `maxFeePerGas`.

On top of that, override the sender's balance to be a large value if
they don't use a paymaster, so that we can produce a gas estimate even
if the sender doesn't currently have enough eth to pay for the gas.