fix(simulation): Reject calls with value if not from sender
The spec says that during validation, calls with value are only allowed
from the account to the entry point, but we incorrectly allowed any
calls with value to the entry point regardless of whom they were from.

By fixing this, we prevent an attack in which unstaked senders could
cause bundles to revert: an earlier sender could call a later sender and
cause it to send some of its eth to the entry point, causing the later
sender to have a lower balance than it did during simulation. The later
sender could then attempt to send to the entry point, see that the call
has failed because it doesn't have enough eth, and branch into
nondeterminism.