fix(estimation): Set zero fees during call gas estimation
When performing call gas estimation, set gas fees to zero so that the
validation doesn't fail for lack of prefund.

This resolves a failure that occurs because of the recent change to
validation call estimation which granted the account a large amount of
eth. That change made it so in cases where the account wasn't funded,
the prefund would succed during validation estimation but fail during
gas estimation.