fix(simulation): don't use abi-generated EntryPoint bytecode
This reverts commit 275414d9268997fa9519cdf77d56f88a202e523d.

We need the v0.6 entry point bytecode so we can use it in state
overrides during call gas estimation. In a previous commit, we started
using bytecode generated by abigen to avoid hardcoding the entry point
bytecode or fetching it from on-chain on every run. For unknown reasons,
the abigen bytecode is faulty and causes reverts with no data, so switch
back to using hardcoded bytecode.